# Remove redundant log output and prometheus metrics update during replay