Before the change:
```
info  2024-11-07T23:18:21.438 nodeos    controller.cpp:1725           replay_block_log     ] 500 of 4983
info  2024-11-07T23:18:21.552 nodeos    controller.cpp:3566           log_applied          ] Received block 733efdcdfaeda3e0... #1000 @ 2024-  11-07T22:42:35.500 signed by eosio [trxs: 0, lib: 999, net: 0, cpu: 100 us, elapsed: 74 us, applying time: 188 us, latency: 2146052 ms]
info  2024-11-07T23:18:21.552 nodeos    controller.cpp:1725           replay_block_log     ] 1000 of 4983
info  2024-11-07T23:18:21.667 nodeos    controller.cpp:1725           replay_block_log     ] 1500 of 4983
info  2024-11-07T23:18:21.781 nodeos    controller.cpp:3566           log_applied          ] Received block d9216020d5998767... #2000 @ 2024-  11-07T22:50:55.500 signed by eosio [trxs: 0, lib: 1999, net: 0, cpu: 100 us, elapsed: 73 us, applying time: 188 us, latency: 1646281 ms]
info  2024-11-07T23:18:21.781 nodeos    controller.cpp:1725           replay_block_log     ] 2000 of 4983
info  2024-11-07T23:18:21.896 nodeos    controller.cpp:1725           replay_block_log     ] 2500 of 4983
info  2024-11-07T23:18:22.011 nodeos    controller.cpp:3566           log_applied          ] Received block 4cea00b8b37f995d... #3000 @ 2024-  11-07T22:59:15.500 signed by eosio [trxs: 0, lib: 2999, net: 0, cpu: 100 us, elapsed: 74 us, applying time: 188 us, latency: 1146511 ms]
info  2024-11-07T23:18:22.011 nodeos    controller.cpp:1725           replay_block_log     ] 3000 of 4983
```

After the change

```
info  2024-11-07T23:31:03.069 nodeos    controller.cpp:1725           replay_block_log     ] 500 of 5227
info  2024-11-07T23:31:03.182 nodeos    controller.cpp:1725           replay_block_log     ] 1000 of 5227
info  2024-11-07T23:31:03.295 nodeos    controller.cpp:1725           replay_block_log     ] 1500 of 5227
info  2024-11-07T23:31:03.408 nodeos    controller.cpp:1725           replay_block_log     ] 2000 of 5227
info  2024-11-07T23:31:03.521 nodeos    controller.cpp:1725           replay_block_log     ] 2500 of 5227
info  2024-11-07T23:31:03.634 nodeos    controller.cpp:1725           replay_block_log     ] 3000 of 5227
```

Resolves https://github.com/AntelopeIO/spring/issues/1017